redirect changes

+++ src/app/shop-cards/shop-cards.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { AgriFreshService } from '../services/agrifresh.service';
 import { Product, CartItem, CartItemInput } from '../model/agrifresh.model';
+import { Router } from "@angular/router";
+import { AuthService } from "../auth/auth.service";
 
 @Component({
   selector: 'app-shop-cards',
@@ -10,7 +12,9 @@ import { Product, CartItem, CartItemInput } from '../model/agrifresh.model';
 export class ShopCardsComponent implements OnInit {
   products: Product[] = [];
   isLoading = false;
-  constructor(private agriFreshService: AgriFreshService) { }
+  constructor(private agriFreshService: AgriFreshService,
+    private authService:AuthService,
+    private router :Router) { }
 
   ngOnInit(): void {
     this.isLoading = true;
@@ -18,9 +22,12 @@ export class ShopCardsComponent implements OnInit {
     this.agriFreshService.getModifiedProductsListener().subscribe((data: { products: Product[], cartItems: CartItem[] }) => {
       if (data && data.products) {
         this.products = data.products;
-        this.isLoading = false;
-        // setTimeout(() => { this.isLoading = false; }, 1009);
+        setTimeout(() => { this.isLoading = false; }, 1009);
       }
+    });
+    this.agriFreshService.getLoadedListener().subscribe(loaded=>{
+      console.log(loaded,"loaded");
+      if(loaded) this.isLoading = false;
     });
   }
 
@@ -29,21 +36,24 @@ export class ShopCardsComponent implements OnInit {
   }
 
   changeQuantity(product: Product, quantity: number) {      
-    // product.cartQuantity=quantity; //optimistic UI
-    if (quantity == 0) {
-      this.agriFreshService.deleteCartItem(product.cartItemId);
-    } else {
-      let cartItemInput: CartItemInput = {
-        itemId: product._id,
-        quantity
-      };
-
-      if (quantity == 1 && !product.cartItemId) {
-        this.agriFreshService.addCartItem(cartItemInput);
-      } else {
-        this.agriFreshService.editCartItem(cartItemInput, product.cartItemId);
-      }
-
+    // product.cartQuantity=quantity; //optimistic UI    
+    if(this.authService.isAuth){
+          if (quantity == 0) {
+            this.agriFreshService.deleteCartItem(product.cartItemId);
+          } else {
+            let cartItemInput: CartItemInput = {
+              itemId: product._id,
+              quantity
+            };
+      
+            if (quantity == 1 && !product.cartItemId) {
+              this.agriFreshService.addCartItem(cartItemInput);
+            } else {
+              this.agriFreshService.editCartItem(cartItemInput, product.cartItemId);
+            }
+        }  
+    }else{
+      this.router.navigate(["/user/login"], { queryParams: { redirect: "shop"}});
     }
   }
